[NWPS-1787] [POC] (1) Leveraging 'com.onehippo.cms7.eforms.hst.util.mail' package from 'com.onehippo.cms7:hippo-addon-eforms-hst' library to improve the code, (2) Code refactoring
