+++ cms-dependencies/pom.xml
@@ -61,6 +61,11 @@
       <groupId>com.onehippo.cms7</groupId>
       <artifactId>hippo-addon-eforms-cms</artifactId>
     </dependency>
+    <!-- Added here to leverage the classes under 'com.onehippo.cms7.eforms.hst.util.mail' package -->
+    <dependency>
+      <groupId>com.onehippo.cms7</groupId>
+      <artifactId>hippo-addon-eforms-hst</artifactId>
+    </dependency>
     <dependency>
       <groupId>com.onehippo.cms7</groupId>
       <artifactId>hippo-addon-urlrewriter-cms-dependencies</artifactId>
