+++ cms/src/main/java/uk/nhs/hee/web/repository/scheduling/NextReviewDateNotification.java
@@ -1,21 +1,17 @@
 package uk.nhs.hee.web.repository.scheduling;
 
-import com.sun.mail.smtp.SMTPMessage;
+import com.onehippo.cms7.eforms.hst.util.mail.*;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.jackrabbit.JcrConstants;
-import org.hippoecm.hst.configuration.hosting.Mount;
-import org.hippoecm.hst.configuration.hosting.VirtualHosts;
-import org.hippoecm.hst.platform.model.HstModel;
-import org.hippoecm.hst.platform.model.HstModelRegistry;
 import org.hippoecm.repository.util.DateTools;
 import org.hippoecm.repository.util.JcrUtils;
-import org.jsoup.Jsoup;
-import org.onehippo.cms7.services.HippoServiceRegistry;
 import org.onehippo.repository.scheduling.RepositoryJob;
 import org.onehippo.repository.scheduling.RepositoryJobExecutionContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import uk.nhs.hee.web.utils.DocumentUtils;
+import uk.nhs.hee.web.utils.MailUtils;
 
 import javax.jcr.Node;
 import javax.jcr.NodeIterator;
@@ -23,20 +19,17 @@ import javax.jcr.RepositoryException;
 import javax.jcr.Session;
 import javax.jcr.query.Query;
 import javax.jcr.query.QueryManager;
-import javax.mail.*;
-import javax.mail.internet.AddressException;
-import javax.mail.internet.InternetAddress;
-import javax.mail.internet.MimeBodyPart;
-import javax.mail.internet.MimeMultipart;
-import javax.naming.Context;
-import javax.naming.InitialContext;
-import javax.naming.NamingException;
-import java.io.UnsupportedEncodingException;
-import java.nio.charset.StandardCharsets;
+import javax.mail.MessagingException;
 import java.text.SimpleDateFormat;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Calendar;
+import java.util.List;
 import java.util.stream.Collectors;
 
+/**
+ * The repository job class which sends email notification with documents to be reviewed in the next 14 days.
+ */
 public class NextReviewDateNotification implements RepositoryJob {
     // Logger
     private static final Logger LOGGER = LoggerFactory.getLogger(NextReviewDateNotification.class);
@@ -69,25 +62,9 @@ public class NextReviewDateNotification implements RepositoryJob {
                     "/*[(%s) and @hippostd:state='published']/element(hee:pageLastNextReview, hee:pageLastNextReview)" +
                     "[@%s and @%s = %s]/../..";
 
-    // CMS context path
-    public static final String CMS_CONTEXT_PATH = "/cms";
-
-    // Host groups
-    public static final String HOST_GROUP_BR_CLOUD = "br-cloud";
-    public static final String HOST_GROUP_DEV_LOCAL_HOST = "dev-localhost";
-
-    // 'brc.environmentname' system property
-    public static final String SYSTEM_PROPERTY_BRC_ENV_NAME = "brc.environmentname";
-
     // Email content/body text placeholders
     public static final String EMAIL_CONTENT_PLACEHOLDER_NEXT_REVIEW_DATE = "{{next_review_date}}";
     public static final String EMAIL_CONTENT_PLACEHOLDER_DOCUMENTS_WITH_LINKS = "{{documents_with_links}}";
-
-    // Mail session bind/lookup name
-    public static final String MAIL_SESSION_NAME = "mail/Session";
-
-    // Regular expression to split comma separated values
-    public static final String SPLIT_COMMA_SEPARATED_VALUES_REGEX = "\\s*,\\s*";
 
     @Override
     public void execute(final RepositoryJobExecutionContext context) throws RepositoryException {
@@ -106,7 +83,7 @@ public class NextReviewDateNotification implements RepositoryJob {
 
                 if (!documentsToBeReviewed.isEmpty()) {
                     // Notify content or channel team
-                    notify(context, session, documentsToBeReviewed);
+                    notify(context, documentsToBeReviewed);
                 } else {
                     LOGGER.info("There are no documents to be reviewed in the next 14 days!");
                 }
@@ -123,6 +100,14 @@ public class NextReviewDateNotification implements RepositoryJob {
         }
     }
 
+    /**
+     * Returns the list of published documents (of {@code documentTypes}) which requires reviewing in the next 14 days.
+     *
+     * @param session       the {@link Session} instance.
+     * @param documentTypes the comma separated list of document types whose documents require reviewing.
+     * @return the list of published documents (of {@code documentTypes}) which requires reviewing in the next 14 days.
+     * @throws RepositoryException thrown when an error occurs during document search.
+     */
     private List<Node> getDocumentsToBeReviewed(final Session session, final String documentTypes)
             throws RepositoryException {
         final QueryManager queryManager = session.getWorkspace().getQueryManager();
@@ -146,7 +131,7 @@ public class NextReviewDateNotification implements RepositoryJob {
     }
 
     private String getFormattedPrimaryDocumentTypes(final String documentTypes) {
-        return Arrays.stream(documentTypes.split(SPLIT_COMMA_SEPARATED_VALUES_REGEX))
+        return uk.nhs.hee.web.utils.StringUtils.transformCommaSeparatedStringIntoStringList(documentTypes)
                 .map(documentType -> String.format("@" + JcrConstants.JCR_PRIMARYTYPE + "='%s'", documentType.trim()))
                 .collect(Collectors.joining(" or "));
     }
@@ -157,9 +142,15 @@ public class NextReviewDateNotification implements RepositoryJob {
         return fourteenDaysFromToday;
     }
 
+    /**
+     * Sends an email notification about the given list of documents to be reviewed {@code documentsToBeReviewed}
+     * based on the configured email specific attributes.
+     *
+     * @param context               the {@link RepositoryJobExecutionContext} instance.
+     * @param documentsToBeReviewed the list of documents to be reviewed.
+     */
     private void notify(
             final RepositoryJobExecutionContext context,
-            final Session session,
             final List<Node> documentsToBeReviewed) {
         try {
             if (!areRequiredEmailAttributesAvailable(context)) {
@@ -176,66 +167,48 @@ public class NextReviewDateNotification implements RepositoryJob {
                     StringUtils.trim(context.getAttribute(ATTRIBUTE_EMAIL_BODY_PLAIN_TEXT_TEMPLATE));
 
             // Create message
-            final javax.mail.Session mailSession = getMailSession();
+            final javax.mail.Session mailSession = MailUtils.getMailSession();
             if (mailSession == null) {
-                LOGGER.warn("Email session '{}' isn't available. Halting the job for now!", MAIL_SESSION_NAME);
+                LOGGER.warn("Email session isn't available. Halting the job for now!");
                 return;
             }
-            final SMTPMessage msg = new SMTPMessage(mailSession);
 
-            // Set from address
-            msg.setFrom(createAddress(fromName, fromEmail));
+            final MailAddress fromMailAddress = new MailAddressImpl(fromName, fromEmail);
+            final MailSender mailSender = new MailSenderImpl(mailSession, fromMailAddress, fromMailAddress);
 
-            // Set sender
-            msg.setSender(createAddress(fromName, fromEmail));
-
-            // Set reply to if needed
-            // msg.setReplyTo();
-
-            // Set envelope from
-            msg.setEnvelopeFrom(fromEmail);
-
-            // Set to/recipient
-            final Address[] addresses = getAddresses(toEmails);
-            msg.addRecipients(Message.RecipientType.TO, addresses);
-
-            // Set subject
-            msg.setSubject(subject);
-
-            // Set content
-            final MimeMultipart mimeAlternativePart = new MimeMultipart("alternative");
             final Pair<String, String> formattedEmailBody =
                     formatEmailBody(bodyHTMLTemplate, bodyPlainTextTemplate, documentsToBeReviewed);
-            LOGGER.debug("Formatted email body HTML: {}", formattedEmailBody.getLeft());
-            LOGGER.debug("Formatted email body plain text: {}", formattedEmailBody.getRight());
 
-            // HTML body
-            final BodyPart htmlBodyPart = new MimeBodyPart();
-            htmlBodyPart.setContent(
+            final MailTemplate mail = new MailTemplate(
+                    mailSender,
+                    MailUtils.getAddresses(toEmails),
+                    subject,
                     formattedEmailBody.getLeft(),
-                    "text/html; charset=" + StandardCharsets.UTF_8.name());
-            mimeAlternativePart.addBodyPart(htmlBodyPart);
+                    formattedEmailBody.getRight());
 
-            // Plain text body
-            final BodyPart textBodyPart = new MimeBodyPart();
-            textBodyPart.setContent(
-                    formattedEmailBody.getRight(),
-                    "text/plain; charset=" + StandardCharsets.UTF_8.name());
-            mimeAlternativePart.addBodyPart(textBodyPart);
-            msg.setContent(mimeAlternativePart);
-
-            // Finally, send
-            Transport.send(msg);
-        } catch(final MessagingException e) {
+            mail.sendMessage();
+        } catch (final MessagingException e) {
             LOGGER.error("Caught error '{}' while preparing and sending the email message", e.getMessage(), e);
         }
     }
 
+    /**
+     * Returns the formatted email body HTML and plain text as a {@link Pair}
+     * by substituting the placeholders (e.g. {{next_review_date}}, {{documents_with_links}}, etc.)
+     * with appropriate values.
+     *
+     * @param emailBodyHTML         the email body HTML which needs to be formatted.
+     * @param emailBodyPlainText    the email body plain text which needs to be formatted.
+     * @param documentsToBeReviewed the {@link List} of documents which needs to be reviewed.
+     * @return the formatted email body HTML and plain text as a {@link Pair}
+     * by substituting the placeholders (e.g. {{next_review_date}}, {{documents_with_links}}, etc.)
+     * with appropriate values.
+     */
     private Pair<String, String> formatEmailBody(
             final String emailBodyHTML,
             final String emailBodyPlainText,
             final List<Node> documentsToBeReviewed) {
-        // Format/update '<next_review_date>' placeholder with 14 days from day in 'd MMMM yyyy' format
+        // Format/update '{{next_review_date}}' placeholder with 14 days from day in 'd MMMM yyyy' format
         final String formatted14DaysFromToday =
                 new SimpleDateFormat("d MMMM yyyy").format(get14DaysFromToday().getTime());
         String formattedEmailBodyHTML = emailBodyHTML.replace(
@@ -246,7 +219,7 @@ public class NextReviewDateNotification implements RepositoryJob {
         // Build documents-to-be-reviewed as a list of links
         final StringBuilder documentsWithLinksForHTMLBody = new StringBuilder("<ul>");
         final StringBuilder documentsWithLinksForPlainTextBody = new StringBuilder();
-        final String cmsBaseContentURL = getCMSBaseContentURL();
+        final String cmsBaseContentURL = DocumentUtils.getDocumentBaseURL();
         for (final Node documentToBeReviewed: documentsToBeReviewed) {
             try {
                 documentsWithLinksForHTMLBody
@@ -272,10 +245,10 @@ public class NextReviewDateNotification implements RepositoryJob {
         // Remove the trailing new line (\n) from the built documents-to-be-reviewed
         if (documentsWithLinksForPlainTextBody.length() >  0 &&
                 documentsWithLinksForPlainTextBody.charAt(documentsWithLinksForPlainTextBody.length() -  1) == '\n') {
-            documentsWithLinksForPlainTextBody.setLength(documentsWithLinksForPlainTextBody.length() -  1);
+            documentsWithLinksForPlainTextBody.setLength(documentsWithLinksForPlainTextBody.length() - 1);
         }
 
-        // Finally, format/update '<documents_with_links>' placeholder with a list of document links to be reviewed.
+        // Finally, format/update '{{documents_with_links}}' placeholder with a list of document links to be reviewed.
         formattedEmailBodyHTML = formattedEmailBodyHTML.replace(
                 EMAIL_CONTENT_PLACEHOLDER_DOCUMENTS_WITH_LINKS, documentsWithLinksForHTMLBody.toString());
         formattedEmailBodyPlainText = formattedEmailBodyPlainText.replace(
@@ -284,13 +257,17 @@ public class NextReviewDateNotification implements RepositoryJob {
         return Pair.of(formattedEmailBodyHTML, formattedEmailBodyPlainText);
     }
 
-    private String convertToEmailPlainBodyText(final String emailBodyHTML) {
-        return Jsoup.parse(emailBodyHTML).wholeText();
-    }
-
+    /**
+     * Returns {@code true} if all the email attributes listed in {@code REQUIRED_EMAIL_ATTRIBUTES} have values.
+     * Otherwise, returns {@code false}.
+     *
+     * @param context the {@link RepositoryJobExecutionContext} instance.
+     * @return {@code true} if all the email attributes listed in {@code REQUIRED_EMAIL_ATTRIBUTES} have values.
+     * Otherwise, returns {@code false}.
+     */
     private boolean areRequiredEmailAttributesAvailable(final RepositoryJobExecutionContext context) {
         final List<String> blankEmailAttributes = new ArrayList<>();
-        for (final String attributeName: REQUIRED_EMAIL_ATTRIBUTES) {
+        for (final String attributeName : REQUIRED_EMAIL_ATTRIBUTES) {
             final String emailAttributeValue = context.getAttribute(attributeName);
             logConfiguredAttributeDebugMsg(attributeName, emailAttributeValue);
 
@@ -310,99 +287,13 @@ public class NextReviewDateNotification implements RepositoryJob {
         return true;
     }
 
-    private Address[] getAddresses(final String toEmails) {
-        return Arrays.stream(toEmails.split(SPLIT_COMMA_SEPARATED_VALUES_REGEX))
-                .map(address -> createAddress(null, address))
-                .toArray(Address[]::new);
-    }
-
-    public static Address createAddress(final String name, final String email) {
-        try {
-            if (StringUtils.isEmpty(name)) {
-                return new InternetAddress(email);
-            }
-
-            return new InternetAddress(email, name);
-        } catch (final UnsupportedEncodingException | AddressException e) {
-            LOGGER.error("Caught error '{}' while creating email address with name='{}' and email='{}'",
-                    e.getMessage(), name, email, e);
-        }
-
-        return null;
-    }
-
-    private javax.mail.Session getMailSession() {
-        Context initialContext = null;
-
-        try {
-            initialContext = new InitialContext();
-            final Context context = (Context) initialContext.lookup("java:comp/env");
-            return (javax.mail.Session) context.lookup(MAIL_SESSION_NAME);
-        } catch (final NamingException e) {
-            LOGGER.error("Error creating email session: {}", MAIL_SESSION_NAME, e);
-            try {
-                if (initialContext != null) {
-                    initialContext.close();
-                }
-            } catch (final NamingException ignore) {
-                // Silently ignore
-            }
-        }
-
-        return null;
-    }
-
+    /**
+     * Logs the given {@code attribute} and its {@code value} as a debug msg.
+     *
+     * @param attribute the attribute name.
+     * @param value     the value of the attribute.
+     */
     private void logConfiguredAttributeDebugMsg(final String attribute, final String value) {
         LOGGER.debug("Configured attribute: {}={}", attribute, value);
-    }
-
-    /**
-     * Returns CMS/Platform base URL for the given {@code hostGroupName} and the mount alias {@code cms-mount}.
-     *
-     * @return the CMS/Platform base URL for the given  {@code hostGroupName} and the mount alias {@code cms-mount}.
-     */
-    private String getCMSBaseContentURL() {
-        String cmsBaseURL = StringUtils.EMPTY;
-
-        final HstModelRegistry hstModelRegistry = HippoServiceRegistry.getService(HstModelRegistry.class);
-        for (final HstModel hstModel : hstModelRegistry.getHstModels()) {
-            try {
-                final VirtualHosts virtualHosts = hstModel.getVirtualHosts();
-                if (CMS_CONTEXT_PATH.equals(virtualHosts.getContextPath())) {
-                    final String hostGroupName =
-                            StringUtils.isNotBlank(System.getProperty(SYSTEM_PROPERTY_BRC_ENV_NAME)) ?
-                                    HOST_GROUP_BR_CLOUD : HOST_GROUP_DEV_LOCAL_HOST;
-                    LOGGER.debug("'{}' virtual host group has been chosen to build the CMS base URL", hostGroupName);
-
-                    final List<Mount> cmsMounts = hstModel.getVirtualHosts().getMountsByHostGroup(hostGroupName);
-                    if (cmsMounts.isEmpty()) {
-                        LOGGER.warn("Host group '{}' is either unavailable in the current instance " +
-                                "or it doesn't have any mounts", hostGroupName);
-                        return StringUtils.EMPTY;
-                    }
-
-                    final Optional<Mount> cmsMountOpt = cmsMounts.stream()
-                            .filter(mount -> mount.isOfType(Mount.LIVE_NAME))
-                            .findFirst();
-                    final Mount cmsMount = cmsMountOpt.orElse(null);
-                    if (cmsMount == null) {
-                        LOGGER.warn("Host group '{}' doesn't seem to have any live mounts", hostGroupName);
-                        return StringUtils.EMPTY;
-                    }
-
-                    cmsBaseURL = cmsMount.getScheme() +
-                            "://" +
-                            cmsMount.getVirtualHost().getHostName() +
-                            (cmsMount.isPortInUrl() ? ":" + cmsMount.getPort() : StringUtils.EMPTY) +
-                            "/cms/content/path";
-
-                    LOGGER.debug("CMS base URL = {}", cmsBaseURL);
-                }
-            } catch (final Exception e) {
-                LOGGER.error("Caught error '{}' while constructing CMS base URL", e.getMessage(), e);
-            }
-        }
-
-        return cmsBaseURL;
     }
 }
