+++ repository-data/application/src/main/resources/hcm-config/configuration/modules/scheduler.yaml
@@ -6,25 +6,24 @@ definitions:
         jcr:primaryType: hipposched:repositoryjob
         hipposched:attributeNames: [documentTypes, emailFromAddress, emailFromName,
           emailToAddresses, emailSubject, emailBodyHTMLTemplate, emailBodyPlainTextTemplate]
-        hipposched:attributeValues: ['hee:blogPost, hee:guidance, hee:news, hee:report,
-            hee:trainingProgrammePage', unknown-email@unknowndomain.com, Do not reply,
-          unknown-email@unknowndomain.com, '[Action required] Documents review', "<!DOCTYPE\
-            \ html>\r\n<html lang=\"en\">\r\n    <head>\r\n        <meta charset=\"\
-            UTF-8\">\r\n        <title>Documents next review date email notification</title>\r\
-            \n    </head>\r\n    <body>\r\n        <p>Dear Content Team,</p>\r\n \
-            \       <p>This is a gentle reminder that the next review dates for the\
-            \ following documents are commencing in 14 days.\r\n            Please\
-            \ make sure to review them on or before {{next_review_date}}.</p>\r\n\
-            \        {{documents_with_links}}\r\n        <p>Thank you,<br>NWP Support</p>\r\
-            \n        <p><small>Note: This is an automated email. Please do not reply,\
-            \ as this mailbox is not monitored.</small></p>\r\n    </body>\r\n</html>",
-          "Dear Content Team,\r\n\r\nThis is a gentle reminder that the next review\
-            \ dates for the following documents are commencing in 14 days. Please\
-            \ make sure to review them on or before {{next_review_date}}.\r\n\r\n\
-            {{documents_with_links}}\r\n\r\nThank you,\r\nNWP Support\r\n\r\nNote:\
-            \ This is an automated email. Please do not reply, as this mailbox is\
-            \ not monitored."]
-        hipposched:enabled: false
+        hipposched:attributeValues: ['hee:blogPost,hee:guidance,hee:news,hee:report,hee:trainingProgrammePage',
+          <<from_email_address>>, Do not reply, <<comma_separated_to_email_addresses>>,
+          '[Action required] Documents review', "<!DOCTYPE html>\r\n<html lang=\"\
+            en\">\r\n    <head>\r\n        <meta charset=\"UTF-8\">\r\n        <title>Documents\
+            \ next review date email notification</title>\r\n    </head>\r\n    <body>\r\
+            \n        <p>Dear Content Team,</p>\r\n        <p>This is a gentle reminder\
+            \ that the next review dates for the following documents are commencing\
+            \ in 14 days.\r\n            Please make sure to review them on or before\
+            \ {{next_review_date}}.</p>\r\n        {{documents_with_links}}\r\n  \
+            \      <p>Thank you,<br>NWP Support</p>\r\n        <p><small>Note: This\
+            \ is an automated email. Please do not reply, as this mailbox is not monitored.</small></p>\r\
+            \n    </body>\r\n</html>", "Dear Content Team,\r\n\r\nThis is a gentle\
+            \ reminder that the next review dates for the following documents are\
+            \ commencing in 14 days. Please make sure to review them on or before\
+            \ {{next_review_date}}.\r\n\r\n{{documents_with_links}}\r\n\r\nThank you,\r\
+            \nNWP Support\r\n\r\nNote: This is an automated email. Please do not reply,\
+            \ as this mailbox is not monitored."]
+        hipposched:enabled: true
         hipposched:repositoryJobClass: uk.nhs.hee.web.repository.scheduling.NextReviewDateNotification
         /hipposched:triggers:
           jcr:primaryType: hipposched:triggers
@@ -32,5 +31,6 @@ definitions:
             jcr:primaryType: hipposched:crontrigger
             jcr:mixinTypes: ['mix:lockable']
             jcr:uuid: 40869f2d-3f45-4710-a125-db5c0d8bad36
-            hipposched:cronExpression: 0 * * * * ?
+            hipposched:cronExpression: 0 0 5 * * ?
             hipposched:enabled: true
+            hipposched:nextFireTime: 2024-02-13T05:00:00Z
